+++ src/parsers/problem/CodeChefNewProblemParser.ts
@@ -5,7 +5,7 @@ import { Parser } from '../Parser';
 
 export class CodeChefNewProblemParser extends Parser {
   public getMatchPatterns(): string[] {
-    return ['https://www.codechef.com/submit-v2/*'];
+    return ['https://www.codechef.com/submit/*'];
   }
 
   public async parse(url: string, html: string): Promise<Sendable> {