+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- Fixed the parsing of CodeChef problems to work with its new URL format
+
 ## [2.34.1](https://github.com/jmerle/competitive-companion/releases/tag/2.34.1) (2022-06-16)
 - Fixed the parsing of Kattis problems and contests to work with Kattis's new design